Bug fixing
Fixed control point bug.
Fixed damage calculation bug.
Fixed shield regen rounding bug.
Fixed shield ability networking bug.

+++ Assets/Scripts/Characters/Character.cs
@@ -55,10 +55,10 @@ public abstract class Character : NetworkBehaviour
 	public CharacterHandler handler;
 
 	public int maxHealth;
-	public int health;
+	public float health;
 
 	public int maxArmour;
-	public int armour;
+	public float armour;
 
     public float armourRegenSpeed;
     public float armourRegenDelay;
